requirements.txt
SpeechRecognition
transformers 
torch 
torchaudio 
pyttsx3 
numpy
pipwin
pyaudio
openai
python-dotenv
pillow
customtkinter 